Add "date" and "organizer email" fields to Meetup
Also add a "province acronym" field to Location

And add two helper functions to print the location in different formats.

Now the locations are shown in the front-end using the DB values.

And the meetup's image is shown in the meetup details page now too 🤙

At last, corrected a typo in a field and did some code clean-up.